requirements.txt
streamlit==1.16.0
transformers
torchaudio==0.13.1
torchvision==0.14.1
spacy
sentencepiece
gTTS==2.3.0
gensim==4.2.0
